refactor: 调整 alignas 的使用使其符合 MSVC 的要求
同时修改部分代码，让其不会在 MSVC 平台下发出警告

+++ src/base.c
@@ -381,7 +381,7 @@ void fklWriteBytevector(const FklBytevector *b, FILE *fp) {
 static char *string_buffer_alloc(void *ptr, size_t len) {
     FklStrBuf *buf = ptr;
     fklStrBufReserve(buf, len + 1);
-    buf->index = len;
+    buf->index = (uint32_t)len;
     char *body = fklStrBufBody(buf);
     body[len] = '\0';
     return body;
